Move setup_logging to helpers.py
This allows to reuse this function in another service.
