+++ scripts/utils/helpers.py
@@ -5,6 +5,8 @@ import re
 import subprocess
 from configparser import ConfigParser
 from itertools import chain
+import logging
+import sys
 
 from tzlocal import get_localzone
 
@@ -101,3 +103,12 @@ def get_wav_files():
     open_recs = get_open_files_in_dir(rec_dir)
     files = [file for file in files if file not in open_recs]
     return files
+
+
+def setup_logging():
+    logger = logging.getLogger()
+    formatter = logging.Formatter("[%(name)s][%(levelname)s] %(message)s")
+    handler = logging.StreamHandler(stream=sys.stdout)
+    handler.setFormatter(formatter)
+    logger.addHandler(handler)
+    logger.setLevel(logging.INFO)
